pyproject.toml
[project]
name = "fast-ima"
version = "0.1.0"
description = "IMA service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
]
